Fix compling for softmax_with_cross_entropy_op.
