+++ paddle/operators/math/math_function_impl.h
@@ -12,6 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
+#pragma once
 #include "paddle/framework/data_type.h"
 #include "paddle/operators/math/math_function.h"
